Move `Widgets.InputWidget` and make it an abstract class
